server/requirements.txt
# Python dependencies for local Whisper transcription
openai-whisper
torch
torchaudio
numpy
ffmpeg-python